+++ etc/visual-studio/libopenthread.vcxproj
@@ -62,6 +62,7 @@
   <ItemGroup>
     <ClCompile Include="..\..\src\core\api\commissioner_api.cpp" />
     <ClCompile Include="..\..\src\core\api\dataset_api.cpp" />
+    <ClCompile Include="..\..\src\core\api\dataset_ftd_api.cpp" />
     <ClCompile Include="..\..\src\core\api\dhcp6_api.cpp" />
     <ClCompile Include="..\..\src\core\api\icmp6_api.cpp" />
     <ClCompile Include="..\..\src\core\api\ip6_api.cpp" />
@@ -73,6 +74,7 @@
     <ClCompile Include="..\..\src\core\api\netdata_api.cpp" />
     <ClCompile Include="..\..\src\core\api\tasklet_api.cpp" />
     <ClCompile Include="..\..\src\core\api\thread_api.cpp" />
+    <ClCompile Include="..\..\src\core\api\thread_ftd_api.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_base.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_client.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_header.cpp" />
