+++ etc/visual-studio/libopenthread.vcxproj.filters
@@ -75,6 +75,9 @@
     <ClCompile Include="..\..\src\core\api\dataset_api.cpp">
       <Filter>Source Files\api</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\core\api\dataset_ftd_api.cpp">
+      <Filter>Source Files\api</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\core\api\dhcp6_api.cpp">
       <Filter>Source Files\api</Filter>
     </ClCompile>
@@ -106,6 +109,9 @@
       <Filter>Source Files\api</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\core\api\thread_api.cpp">
+      <Filter>Source Files\api</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\core\api\thread_ftd_api.cpp">
       <Filter>Source Files\api</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\core\coap\coap_base.cpp">
