+++ etc/visual-studio/libopenthread_k.vcxproj
@@ -69,6 +69,7 @@
   <ItemGroup>
     <ClCompile Include="..\..\src\core\api\commissioner_api.cpp" />
     <ClCompile Include="..\..\src\core\api\dataset_api.cpp" />
+    <ClCompile Include="..\..\src\core\api\dataset_ftd_api.cpp" />
     <ClCompile Include="..\..\src\core\api\dhcp6_api.cpp" />
     <ClCompile Include="..\..\src\core\api\icmp6_api.cpp" />
     <ClCompile Include="..\..\src\core\api\ip6_api.cpp" />
@@ -79,6 +80,7 @@
     <ClCompile Include="..\..\src\core\api\netdata_api.cpp" />
     <ClCompile Include="..\..\src\core\api\tasklet_api.cpp" />
     <ClCompile Include="..\..\src\core\api\thread_api.cpp" />
+    <ClCompile Include="..\..\src\core\api\thread_ftd_api.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_base.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_client.cpp" />
     <ClCompile Include="..\..\src\core\coap\coap_header.cpp" />
@@ -152,6 +154,7 @@
     <ClInclude Include="..\..\include\openthread\commissioner.h" />
     <ClInclude Include="..\..\include\openthread\crypto.h" />
     <ClInclude Include="..\..\include\openthread\dataset.h" />
+    <ClInclude Include="..\..\include\openthread\dataset_ftd.h" />
     <ClInclude Include="..\..\include\openthread\dhcp6_client.h" />
     <ClInclude Include="..\..\include\openthread\dhcp6_server.h" />
     <ClInclude Include="..\..\include\openthread\icmp6.h" />
@@ -165,6 +168,7 @@
     <ClInclude Include="..\..\include\openthread\openthread.h" />
     <ClInclude Include="..\..\include\openthread\tasklet.h" />
     <ClInclude Include="..\..\include\openthread\thread.h" />
+    <ClInclude Include="..\..\include\openthread\thread_ftd.h" />
     <ClInclude Include="..\..\include\openthread\types.h" />
     <ClInclude Include="..\..\include\openthread-windows-config.h" />
     <ClInclude Include="..\..\include\openthread\udp.h" />
