+++ examples/drivers/windows/otApi/precomp.h
@@ -22,6 +22,8 @@
 #define OTAPI EXTERN_C __declspec(dllexport)
 
 #include <openthread/openthread.h>
+#include <openthread/dataset_ftd.h>
+#include <openthread/thread_ftd.h>
 #include <openthread/commissioner.h>
 #include <openthread/joiner.h>
 #include <openthread/platform/logging-windows.h>
