+++ examples/drivers/windows/otNodeApi/precomp.h
@@ -50,6 +50,8 @@ using namespace std;
 #define OTNODEAPI EXTERN_C __declspec(dllexport)
 
 #include <openthread/openthread.h>
+#include <openthread/dataset_ftd.h>
+#include <openthread/thread_ftd.h>
 #include <openthread/commissioner.h>
 #include <openthread/joiner.h>
 #include <openthread/platform/logging-windows.h>
