+++ include/openthread/Makefile.am
@@ -52,6 +52,7 @@ openthread_headers                      = \
     commissioner.h                        \
     crypto.h                              \
     dataset.h                             \
+    dataset_ftd.h                         \
     diag.h                                \
     dhcp6_client.h                        \
     dhcp6_server.h                        \
@@ -69,6 +70,7 @@ openthread_headers                      = \
     openthread.h                          \
     tasklet.h                             \
     thread.h                              \
+    thread_ftd.h                          \
     types.h                               \
     udp.h                                 \
     $(NULL)
