+++ include/openthread/dataset.h
@@ -29,7 +29,7 @@
 /**
  * @file
  * @brief
- *  This file defines the OpenThread Operational Dataset API.
+ *  This file defines the OpenThread Operational Dataset API (for both FTD and MTD).
  */
 
 #ifndef OPENTHREAD_DATASET_H_
@@ -74,19 +74,6 @@ OTAPI bool OTCALL otDatasetIsCommissioned(otInstance *aInstance);
 OTAPI ThreadError OTCALL otDatasetGetActive(otInstance *aInstance, otOperationalDataset *aDataset);
 
 /**
- * This function sets the Active Operational Dataset.
- *
- * @param[in]  aInstance A pointer to an OpenThread instance.
- * @param[in]  aDataset  A pointer to the Active Operational Dataset.
- *
- * @retval kThreadError_None         Successfully set the Active Operational Dataset.
- * @retval kThreadError_NoBufs       Insufficient buffer space to set the Active Operational Datset.
- * @retval kThreadError_InvalidArgs  @p aDataset was NULL.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSetActive(otInstance *aInstance, const otOperationalDataset *aDataset);
-
-/**
  * This function gets the Pending Operational Dataset.
  *
  * @param[in]   aInstance A pointer to an OpenThread instance.
@@ -97,101 +84,6 @@ OTAPI ThreadError OTCALL otDatasetSetActive(otInstance *aInstance, const otOpera
  *
  */
 OTAPI ThreadError OTCALL otDatasetGetPending(otInstance *aInstance, otOperationalDataset *aDataset);
-
-/**
- * This function sets the Pending Operational Dataset.
- *
- * @param[in]  aInstance A pointer to an OpenThread instance.
- * @param[in]  aDataset  A pointer to the Pending Operational Dataset.
- *
- * @retval kThreadError_None         Successfully set the Pending Operational Dataset.
- * @retval kThreadError_NoBufs       Insufficient buffer space to set the Pending Operational Dataset.
- * @retval kThreadError_InvalidArgs  @p aDataset was NULL.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSetPending(otInstance *aInstance, const otOperationalDataset *aDataset);
-
-/**
- * This function sends MGMT_ACTIVE_GET.
- *
- * @param[in]  aInstance  A pointer to an OpenThread instance.
- * @param[in]  aTlvTypes  A pointer to the TLV Types.
- * @param[in]  aLength    The length of TLV Types.
- * @param[in]  aAddress   A pointer to the IPv6 destination, if it is NULL, will use Leader ALOC as default.
- *
- * @retval kThreadError_None         Successfully send the meshcop dataset command.
- * @retval kThreadError_NoBufs       Insufficient buffer space to send.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSendMgmtActiveGet(otInstance *aInstance, const uint8_t *aTlvTypes, uint8_t aLength,
-                                                    const otIp6Address *aAddress);
-
-/**
- * This function sends MGMT_ACTIVE_SET.
- *
- * @param[in]  aInstance  A pointer to an OpenThread instance.
- * @param[in]  aDataset   A pointer to operational dataset.
- * @param[in]  aTlvs      A pointer to TLVs.
- * @param[in]  aLength    The length of TLVs.
- *
- * @retval kThreadError_None         Successfully send the meshcop dataset command.
- * @retval kThreadError_NoBufs       Insufficient buffer space to send.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSendMgmtActiveSet(otInstance *aInstance, const otOperationalDataset *aDataset,
-                                                    const uint8_t *aTlvs, uint8_t aLength);
-
-/**
- * This function sends MGMT_PENDING_GET.
- *
- * @param[in]  aInstance  A pointer to an OpenThread instance.
- * @param[in]  aTlvTypes  A pointer to the TLV Types.
- * @param[in]  aLength    The length of TLV Types.
- * @param[in]  aAddress   A pointer to the IPv6 destination, if it is NULL, will use Leader ALOC as default.
- *
- * @retval kThreadError_None         Successfully send the meshcop dataset command.
- * @retval kThreadError_NoBufs       Insufficient buffer space to send.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSendMgmtPendingGet(otInstance *aInstance, const uint8_t *aTlvTypes, uint8_t aLength,
-                                                     const otIp6Address *aAddress);
-
-/**
- * This function sends MGMT_PENDING_SET.
- *
- * @param[in]  aInstance  A pointer to an OpenThread instance.
- * @param[in]  aDataset   A pointer to operational dataset.
- * @param[in]  aTlvs      A pointer to TLVs.
- * @param[in]  aLength    The length of TLVs.
- *
- * @retval kThreadError_None         Successfully send the meshcop dataset command.
- * @retval kThreadError_NoBufs       Insufficient buffer space to send.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSendMgmtPendingSet(otInstance *aInstance, const otOperationalDataset *aDataset,
-                                                     const uint8_t *aTlvs, uint8_t aLength);
-
-/**
- * Get minimal delay timer.
- *
- * @param[in]  aInstance A pointer to an OpenThread instance.
- *
- * @retval the value of minimal delay timer (in ms).
- *
- */
-OTAPI uint32_t OTCALL otDatasetGetDelayTimerMinimal(otInstance *aInstance);
-
-/**
- * Set minimal delay timer.
- *
- * @param[in]  aInstance           A pointer to an OpenThread instance.
- * @param[in]  aDelayTimerMinimal  The value of minimal delay timer (in ms).
- *
- * @retval  kThreadError_None         Successfully set minimal delay timer.
- * @retval  kThreadError_InvalidArgs  If @p aDelayTimerMinimal is not valid.
- *
- */
-OTAPI ThreadError OTCALL otDatasetSetDelayTimerMinimal(otInstance *aInstance, uint32_t aDelayTimerMinimal);
 
 /**
  * @}
