+++ src/core/api/dataset_api.cpp
@@ -28,7 +28,7 @@
 
 /**
  * @file
- *   This file implements the OpenThread Operational Dataset API.
+ *   This file implements the OpenThread Operational Dataset API (for both FTD and MTD).
  */
 
 #include "openthread/dataset.h"
@@ -64,18 +64,6 @@ exit:
     return error;
 }
 
-ThreadError otDatasetSetActive(otInstance *aInstance, const otOperationalDataset *aDataset)
-{
-    ThreadError error;
-
-    VerifyOrExit(aDataset != NULL, error = kThreadError_InvalidArgs);
-
-    error = aInstance->mThreadNetif.GetActiveDataset().Set(*aDataset);
-
-exit:
-    return error;
-}
-
 ThreadError otDatasetGetPending(otInstance *aInstance, otOperationalDataset *aDataset)
 {
     ThreadError error = kThreadError_None;
@@ -86,50 +74,4 @@ ThreadError otDatasetGetPending(otInstance *aInstance, otOperationalDataset *aDa
 
 exit:
     return error;
-}
-
-ThreadError otDatasetSetPending(otInstance *aInstance, const otOperationalDataset *aDataset)
-{
-    ThreadError error;
-
-    VerifyOrExit(aDataset != NULL, error = kThreadError_InvalidArgs);
-
-    error = aInstance->mThreadNetif.GetPendingDataset().Set(*aDataset);
-
-exit:
-    return error;
-}
-
-ThreadError otDatasetSendMgmtActiveGet(otInstance *aInstance, const uint8_t *aTlvTypes, uint8_t aLength,
-                                       const otIp6Address *aAddress)
-{
-    return aInstance->mThreadNetif.GetActiveDataset().SendGetRequest(aTlvTypes, aLength, aAddress);
-}
-
-ThreadError otDatasetSendMgmtActiveSet(otInstance *aInstance, const otOperationalDataset *aDataset,
-                                       const uint8_t *aTlvs, uint8_t aLength)
-{
-    return aInstance->mThreadNetif.GetActiveDataset().SendSetRequest(*aDataset, aTlvs, aLength);
-}
-
-ThreadError otDatasetSendMgmtPendingGet(otInstance *aInstance, const uint8_t *aTlvTypes, uint8_t aLength,
-                                        const otIp6Address *aAddress)
-{
-    return aInstance->mThreadNetif.GetPendingDataset().SendGetRequest(aTlvTypes, aLength, aAddress);
-}
-
-ThreadError otDatasetSendMgmtPendingSet(otInstance *aInstance, const otOperationalDataset *aDataset,
-                                        const uint8_t *aTlvs, uint8_t aLength)
-{
-    return aInstance->mThreadNetif.GetPendingDataset().SendSetRequest(*aDataset, aTlvs, aLength);
-}
-
-uint32_t otDatasetGetDelayTimerMinimal(otInstance *aInstance)
-{
-    return aInstance->mThreadNetif.GetLeader().GetDelayTimerMinimal();
-}
-
-ThreadError otDatasetSetDelayTimerMinimal(otInstance *aInstance, uint32_t aDelayTimerMinimal)
-{
-    return aInstance->mThreadNetif.GetLeader().SetDelayTimerMinimal(aDelayTimerMinimal);
 }
