+++ src/core/meshcop/dataset_manager.cpp
@@ -156,6 +156,7 @@ ThreadError DatasetManager::ApplyConfiguration(void)
     return error;
 }
 
+#if OPENTHREAD_FTD
 ThreadError DatasetManager::Set(const otOperationalDataset &aDataset, uint8_t &aFlags)
 {
     ThreadError error = kThreadError_None;
@@ -185,6 +186,7 @@ ThreadError DatasetManager::Set(const otOperationalDataset &aDataset, uint8_t &a
 exit:
     return error;
 }
+#endif // OPENTHREAD_FTD
 
 ThreadError DatasetManager::Clear(uint8_t &aFlags, bool aOnlyClearNetwork)
 {
@@ -342,6 +344,7 @@ void DatasetManager::Get(Coap::Header &aHeader, Message &aMessage, const Ip6::Me
     SendGetResponse(aHeader, aMessageInfo, tlvs, length);
 }
 
+#if OPENTHREAD_FTD
 ThreadError DatasetManager::Set(Coap::Header &aHeader, Message &aMessage, const Ip6::MessageInfo &aMessageInfo)
 {
     Tlv tlv;
@@ -813,6 +816,7 @@ exit:
         message->Free();
     }
 }
+#endif // OPENTHREAD_FTD
 
 void DatasetManager::SendGetResponse(const Coap::Header &aRequestHeader, const Ip6::MessageInfo &aMessageInfo,
                                      uint8_t *aTlvs, uint8_t aLength)
@@ -910,6 +914,7 @@ exit:
     return error;
 }
 
+#if OPENTHREAD_FTD
 ThreadError ActiveDatasetBase::Set(const otOperationalDataset &aDataset)
 {
     ThreadError error = kThreadError_None;
@@ -921,6 +926,7 @@ ThreadError ActiveDatasetBase::Set(const otOperationalDataset &aDataset)
 exit:
     return error;
 }
+#endif // OPENTHREAD_FTD
 
 ThreadError ActiveDatasetBase::Set(const Dataset &aDataset)
 {
@@ -999,6 +1005,7 @@ exit:
     return error;
 }
 
+#if OPENTHREAD_FTD
 ThreadError PendingDatasetBase::Set(const otOperationalDataset &aDataset)
 {
     ThreadError error = kThreadError_None;
@@ -1010,6 +1017,7 @@ ThreadError PendingDatasetBase::Set(const otOperationalDataset &aDataset)
 exit:
     return error;
 }
+#endif // OPENTHREAD_FTD
 
 ThreadError PendingDatasetBase::Set(const Dataset &aDataset)
 {
@@ -1118,45 +1126,6 @@ void PendingDatasetBase::HandleTimer(void)
     mNetif.GetActiveDataset().Set(mNetwork);
 
     Clear(false);
-}
-
-void PendingDatasetBase::ApplyActiveDataset(const Timestamp &aTimestamp, Message &aMessage)
-{
-    uint16_t offset = aMessage.GetOffset();
-    DelayTimerTlv delayTimer;
-    uint8_t flags;
-
-    VerifyOrExit(mNetif.GetMle().IsAttached());
-
-    while (offset < aMessage.GetLength())
-    {
-        OT_TOOL_PACKED_BEGIN
-        struct
-        {
-            Tlv tlv;
-            uint8_t value[Dataset::kMaxValueSize];
-        } OT_TOOL_PACKED_END data;
-
-        aMessage.Read(offset, sizeof(Tlv), &data.tlv);
-        aMessage.Read(offset + sizeof(Tlv), data.tlv.GetLength(), data.value);
-        mNetwork.Set(data.tlv);
-        offset += sizeof(Tlv) + data.tlv.GetLength();
-    }
-
-    // add delay timer tlv
-    delayTimer.Init();
-    delayTimer.SetDelayTimer(mNetif.GetLeader().GetDelayTimerMinimal());
-    mNetwork.Set(delayTimer);
-
-    // add pending timestamp tlv
-    mNetwork.SetTimestamp(aTimestamp);
-    HandleNetworkUpdate(flags);
-
-    // reset delay timer
-    ResetDelayTimer(kFlagNetworkUpdated);
-
-exit:
-    return;
 }
 
 void PendingDatasetBase::HandleNetworkUpdate(uint8_t &aFlags)
